OPSMN-1860 : Allowing specimen quantity & conc notation in the form 10e6
Description :
- User is allowed to enter concentration and quantity fields on specimen pages - add, edit, derivative, aliquot, BO and
  Specimen requirement
- In specimen concentration and quantity field, any number entered more than 1 million is converted into scientific notation
- In specimen edit, overview, the field value is be displayed in scientific notations.

+++ WEB-INF/src/com/krishagni/catissueplus/core/biospecimen/events/SpecimenRequirementDetail.java
@@ -36,7 +36,7 @@ public class SpecimenRequirementDetail implements Comparable<SpecimenRequirement
 	
 	private BigDecimal initialQty;
 	
-	private Double concentration;
+	private BigDecimal concentration;
 	
 	private UserSummary collector;
 	
@@ -142,11 +142,11 @@ public class SpecimenRequirementDetail implements Comparable<SpecimenRequirement
 		this.initialQty = initialQty;
 	}
 
-	public Double getConcentration() {
+	public BigDecimal getConcentration() {
 		return concentration;
 	}
 
-	public void setConcentration(Double concentration) {
+	public void setConcentration(BigDecimal concentration) {
 		this.concentration = concentration;
 	}
